% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetVisiumData.R
\name{GetVisiumData}
\alias{GetVisiumData}
\title{Load Example Visium Spatial Transcriptomics Data}
\usage{
GetVisiumData()
}
\value{
A Seurat object containing spatial transcriptomics data.
}
\description{
Loads a pre-processed Seurat object containing 10X Visium spatial
transcriptomics data, for use in tutorials and examples.
}
\examples{
\dontrun{
spat <- GetVisiumData()
}

}
